backend/user-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
elastic-apm
python-dotenv